Complete `Vocabularies` async port.
- Moved to `vocabularies_async` and `types` to be consistent with `dev_async`.
- Set `internal_resource` to the old resource.
- Added `ResourceConfig`.
- Added module to `default_settings`.
- Addressed feedbacks.
