+++ superdesk/vocabularies/vocabularies.py
@@ -51,6 +51,7 @@ vocab_schema = {
 
 
 class VocabulariesResource(Resource):
+    internal_resource = True
     schema = {
         "_id": {
             "type": "string",
